// src/interfaces/Addresses.tsx
// Generated by https://quicktype.io

export interface Address {
    cantidad:    number;
    direcciones: Direcciones[];
    inicio:      number;
    parametros:  Parametros;
    total:       number;
}

export interface Direcciones {
    altura:           DireccionesAltura;
    calle:            Calle;
    calle_cruce_1:    Calle;
    calle_cruce_2:    Calle;
    departamento:     Departamento;
    localidad_censal: Departamento;
    nomenclatura:     string;
    piso:             null;
    provincia:        Departamento;
    ubicacion:        Ubicacion;
}

export interface DireccionesAltura {
    unidad: null;
    valor:  number;
}

export interface Calle {
    categoria: null | string;
    id:        null | string;
    nombre:    null | string;
}

export interface Departamento {
    id:     string;
    nombre: string;
}

export interface Ubicacion {
    lat: number;
    lon: number;
}

export interface Parametros {
    departamento: string;
    direccion:    Direccion;
}

export interface Direccion {
    altura: DireccionAltura;
    calles: string[];
    piso:   null;
    tipo:   string;
}

export interface DireccionAltura {
    unidad: null;
    valor:  string;
}